rename IntersectionSelected
