+++ server/src/test/java/com/mankomania/game/server/data/TestServerDataBaseTurn.java
@@ -4,7 +4,7 @@ import com.badlogic.gdx.utils.IntArray;
 import com.esotericsoftware.kryonet.Connection;
 import com.esotericsoftware.kryonet.Server;
 import com.mankomania.game.core.network.messages.clienttoserver.baseturn.DiceResultMessage;
-import com.mankomania.game.core.network.messages.clienttoserver.baseturn.IntersectionSelectedMessage;
+import com.mankomania.game.core.network.messages.clienttoserver.baseturn.IntersectionSelection;
 import com.mankomania.game.core.network.messages.servertoclient.baseturn.PlayerCanRollDiceMessage;
 import com.mankomania.game.core.network.messages.servertoclient.baseturn.PlayerMoves;
 import org.junit.jupiter.api.Assertions;
@@ -270,7 +270,7 @@ public class TestServerDataBaseTurn {
         // set game state to something that should not work
         this.serverData.setCurrentState(GameState.WAIT_FOR_DICE_RESULT);
         // call method while in wrong state
-        this.serverData.gotIntersectionSelectionMessage(new IntersectionSelectedMessage(), 123);
+        this.serverData.gotIntersectionSelectionMessage(new IntersectionSelection(), 123);
 
         // check if none of the send methods of the server were called
         verify(this.mockedServer, times(0)).sendToAllTCP(Mockito.any());
@@ -286,10 +286,10 @@ public class TestServerDataBaseTurn {
         // set game state so it works
         this.serverData.setCurrentState(GameState.WAIT_INTERSECTION_SELECTION);
         // create a message that the function should handle, but using a different connection id
-        IntersectionSelectedMessage intersectionSelectedMessage = new IntersectionSelectedMessage();
-        intersectionSelectedMessage.setPlayerIndex(1);
-        intersectionSelectedMessage.setFieldIndex(13);
-        this.serverData.gotIntersectionSelectionMessage(intersectionSelectedMessage, 34);
+        IntersectionSelection intersectionSelection = new IntersectionSelection();
+        intersectionSelection.setPlayerIndex(1);
+        intersectionSelection.setFieldIndex(13);
+        this.serverData.gotIntersectionSelectionMessage(intersectionSelection, 34);
 
         // check if none of the send methods of the server were called
         verify(this.mockedServer, times(0)).sendToAllTCP(Mockito.any());
